Create updated-movement-beta2.circ
@Xenearsa
1. Created the new even better ball movement.

First we calculate its position on an 256x256 matrix — then draw it on our 32X32 screen.

The speed of the ball can be any number (but at very high speeds it doesn't look good).
Some basic physics simulation is implemented. When the ball reaches a wall/racket, there's a slight "bounce" delay like in real life.